+++ packages/components/badge/stories/EntityStatusBadge.stories.tsx
@@ -38,44 +38,67 @@ export const Overview = () => (
         <EntityStatusBadge entityStatus="published" />
       </Flex>
       <Flex marginRight="spacingS">
+        <EntityStatusBadge entityStatus="published" isScheduled />
+      </Flex>
+      <Flex marginRight="spacingS">
         <EntityStatusBadge entityStatus="published" size="small" />
       </Flex>
     </Flex>
+
     <Flex marginBottom="spacingM" alignItems="center">
       <Flex marginRight="spacingS">
         <EntityStatusBadge entityStatus="new" />
       </Flex>
       <Flex marginRight="spacingS">
+        <EntityStatusBadge entityStatus="new" isScheduled />
+      </Flex>
+      <Flex marginRight="spacingS">
         <EntityStatusBadge entityStatus="new" size="small" />
       </Flex>
     </Flex>
+
     <Flex marginBottom="spacingM" alignItems="center">
       <Flex marginRight="spacingS">
         <EntityStatusBadge entityStatus="changed" />
       </Flex>
       <Flex marginRight="spacingS">
+        <EntityStatusBadge entityStatus="changed" isScheduled />
+      </Flex>
+      <Flex marginRight="spacingS">
         <EntityStatusBadge entityStatus="changed" size="small" />
       </Flex>
     </Flex>
+
     <Flex marginBottom="spacingM" alignItems="center">
       <Flex marginRight="spacingS">
         <EntityStatusBadge entityStatus="deleted" />
       </Flex>
       <Flex marginRight="spacingS">
+        <EntityStatusBadge entityStatus="deleted" isScheduled />
+      </Flex>
+      <Flex marginRight="spacingS">
         <EntityStatusBadge entityStatus="deleted" size="small" />
       </Flex>
     </Flex>
+
     <Flex marginBottom="spacingM" alignItems="center">
       <Flex marginRight="spacingS">
         <EntityStatusBadge entityStatus="draft" />
       </Flex>
       <Flex marginRight="spacingS">
+        <EntityStatusBadge entityStatus="draft" isScheduled />
+      </Flex>
+      <Flex marginRight="spacingS">
         <EntityStatusBadge entityStatus="draft" size="small" />
       </Flex>
     </Flex>
+
     <Flex marginBottom="spacingM" alignItems="center">
       <Flex marginRight="spacingS">
         <EntityStatusBadge entityStatus="archived" />
+      </Flex>
+      <Flex marginRight="spacingS">
+        <EntityStatusBadge entityStatus="archived" isScheduled />
       </Flex>
       <Flex marginRight="spacingS">
         <EntityStatusBadge entityStatus="archived" size="small" />