+++ packages/components/button/src/Button/Button.styles.ts
@@ -2,6 +2,7 @@ import { css } from 'emotion';
 import type { CSSObject } from '@emotion/serialize';
 import tokens from '@contentful/f36-tokens';
 import { ButtonSize, ButtonVariant, ButtonStylesProps } from '../types';
+import { hexToRGBA } from '@contentful/f36-utils';
 
 const variantActiveStyles = (variant: ButtonVariant): CSSObject => {
   switch (variant) {
@@ -14,7 +15,7 @@ const variantActiveStyles = (variant: ButtonVariant): CSSObject => {
     case 'negative':
       return { backgroundColor: tokens.red800, borderColor: tokens.red800 };
     case 'transparent':
-      return { backgroundColor: tokens.gray100, borderColor: tokens.gray100 };
+      return { backgroundColor: hexToRGBA(tokens.gray900, 0.1) };
     default:
       return {};
   }
@@ -116,17 +117,15 @@ const variantToStyles = (variant: ButtonVariant): CSSObject => {
       };
     case 'transparent':
       return {
-        color: tokens.gray800,
+        color: tokens.gray900,
         background: 'none',
         borderColor: 'transparent',
         boxShadow: 'none',
         '&:hover': {
-          backgroundColor: tokens.gray100,
-          color: tokens.gray900,
+          backgroundColor: hexToRGBA(tokens.gray900, 0.05),
         },
         '&:active': variantActiveStyles(variant),
         '&:focus': {
-          backgroundColor: 'transparent',
           boxShadow: tokens.glowPrimary,
         },
         '&:focus:not(:focus-visible)': {
@@ -224,7 +223,7 @@ export const getStyles = () => ({
       ...(isActive
         ? {
             transition: 'none',
-            '&, &:hover': variantActiveStyles(variant),
+            '&, &:focus': variantActiveStyles(variant),
           }
         : {}),
     }),