+++ packages/components/card/examples/Card/WithCustomDragHandle.tsx
@@ -1,76 +1,88 @@
 import React from 'react';
 import { Box, Card, DragHandle, Flex } from '@contentful/f36-components';
 import { css } from 'emotion';
+import { DndContext } from '@dnd-kit/core';
 import {
-  SortableContainer,
-  SortableElement,
-  SortableHandle,
-} from 'react-sortable-hoc';
-import { arrayMoveImmutable } from 'array-move';
+  arrayMove,
+  verticalListSortingStrategy,
+  SortableContext,
+  useSortable,
+} from '@dnd-kit/sortable';
+import { CSS } from '@dnd-kit/utilities';
 
 export default function WithCustomDragHandle() {
   const styles = {
+    card: css({
+      // This lets us change z-index when dragging
+      position: 'relative',
+    }),
     dragHandle: css({
       alignSelf: 'stretch',
     }),
   };
-  const [items, updateItems] = React.useState([
+  const [items, setItems] = React.useState([
     'Tech',
     'News',
     'CMS',
     'Contentful',
   ]);
 
-  const swapItems = React.useCallback(
-    ({ oldIndex, newIndex }) => {
-      // `arrayMoveImmutable` is imported from 'array-move'
-      const newItems = arrayMoveImmutable(items, oldIndex, newIndex);
-      updateItems(newItems);
-    },
-    [items],
-  );
+  function SortableCard({ id }) {
+    const { attributes, listeners, setNodeRef, transform, transition, active } =
+      useSortable({
+        id,
+      });
+    const zIndex = active && active.id === id ? 1 : 0;
+    const style = {
+      transform: CSS.Translate.toString(transform),
+      transition,
+      zIndex,
+    };
 
-  const SortableList = SortableContainer((props) => (
-    <Flex flexDirection="column">{props.children}</Flex>
-  ));
-
-  const SortableDragHandle = SortableHandle(() => (
-    <DragHandle as="button" className={styles.dragHandle} label="Move card" />
-  ));
-
-  const SortableCard = SortableElement(({ label }) => {
     return (
       <Card
-        dragHandleRender={() => <SortableDragHandle />}
+        className={styles.card}
+        dragHandleRender={() => (
+          <DragHandle
+            as="button"
+            className={styles.dragHandle}
+            label="Move card"
+            {...attributes}
+            {...listeners}
+          />
+        )}
         padding="none"
         withDragHandle
+        ref={setNodeRef}
+        style={style}
       >
-        <Box padding="spacingM">{label}</Box>
+        <Box padding="spacingM">{id}</Box>
       </Card>
     );
-  });
+  }
+
+  const handleDragEnd = (event) => {
+    const { active, over } = event;
+
+    if (active && over && active.id !== over.id) {
+      setItems((items) => {
+        const oldIndex = items.indexOf(active.id);
+        const newIndex = items.indexOf(over.id);
+
+        return arrayMove(items, oldIndex, newIndex);
+      });
+    }
+  };
 
   return (
-    <SortableList
-      useDragHandle
-      axis="y"
-      distance={10}
-      onSortEnd={({ oldIndex, newIndex }) => {
-        swapItems({ oldIndex, newIndex });
-      }}
-      shouldCancelStart={(event) => {
-        const interactiveElements = ['INPUT', 'TEXTAREA', 'SELECT', 'OPTION'];
-
-        if (interactiveElements.indexOf(event.target.tagName) !== -1) {
-          return true;
-        }
-
-        return false;
-      }}
-    >
-      {items.map((item, index) => (
-        <SortableCard key={item} index={index} label={item} />
-      ))}
-    </SortableList>
+    <DndContext onDragEnd={handleDragEnd}>
+      <SortableContext items={items} strategy={verticalListSortingStrategy}>
+        <Flex flexDirection="column" gap="spacingS">
+          {items.map((item) => (
+            <SortableCard key={item} id={item} />
+          ))}
+        </Flex>
+      </SortableContext>
+    </DndContext>
   );
 }