+++ packages/forma-36-react-components/src/index.ts
@@ -11,6 +11,7 @@ export * from '@contentful/f36-datetime';
 export * from '@contentful/f36-datepicker';
 export * from '@contentful/f36-drag-handle';
 export * from '@contentful/f36-entity-list';
+export * from '@contentful/f36-empty-state';
 export * from '@contentful/f36-forms';
 export * from '@contentful/f36-icon';
 export * from '@contentful/f36-list';