+++ packages/website/components/LiveEditor/ComponentSource.tsx
@@ -11,6 +11,7 @@ import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live';
 import tokens from '@contentful/f36-tokens';
 import * as f36Components from '@contentful/f36-components';
 import { Multiselect } from '@contentful/f36-multiselect';
+import { NavList } from '@contentful/f36-navlist';
 import * as f36utils from '@contentful/f36-utils';
 import { useForm, useController } from 'react-hook-form';
 import { MdAccessAlarm } from 'react-icons/md';
@@ -21,12 +22,15 @@ import { theme } from './theme';
 import { formatSourceCode } from './utils';
 import * as coder from '../../utils/coder';
 import FocusLock from 'react-focus-lock';
+import { DndContext } from '@dnd-kit/core';
 import {
-  SortableContainer,
-  SortableElement,
-  SortableHandle,
-} from 'react-sortable-hoc';
-import { arrayMoveImmutable } from 'array-move';
+  arrayMove,
+  horizontalListSortingStrategy,
+  SortableContext,
+  useSortable,
+  verticalListSortingStrategy,
+} from '@dnd-kit/sortable';
+import { CSS } from '@dnd-kit/utilities';
 import { format, parse, isValid } from 'date-fns';
 
 const liveProviderScope = {
@@ -34,6 +38,7 @@ const liveProviderScope = {
   ...f36Components,
   ...f36utils,
   Multiselect, // Remove when added to f36-components
+  NavList, // Remove when added to f36-components
   css,
   f36icons,
   tokens,
@@ -44,15 +49,21 @@ const liveProviderScope = {
   useRef,
   useMemo,
   useContext,
-  // other
+  // react-hook-form
   useForm,
   useController,
+  // other
   MdAccessAlarm,
   FocusLock,
-  SortableContainer,
-  SortableElement,
-  SortableHandle,
-  arrayMoveImmutable,
+  // dnd-kit
+  arrayMove,
+  CSS,
+  DndContext,
+  horizontalListSortingStrategy,
+  SortableContext,
+  useSortable,
+  verticalListSortingStrategy,
+  // date-fns
   format,
   parse,
   isValid,