+++ packages/website/next.config.js
@@ -14,4 +14,14 @@ module.exports = withMDX({
       },
     ];
   },
+  images: {
+    remotePatterns: [
+      {
+        protocol: 'https',
+        hostname: 'images.ctfassets.net',
+        port: '',
+        pathname: '/**',
+      },
+    ],
+  },
 });