+++ packages/components/table/CHANGELOG.md
@@ -1,5 +1,139 @@
 # Change Log
 
+## 4.41.1
+
+### Patch Changes
+
+- [#2495](https://github.com/contentful/forma-36/pull/2495) [`bb6b97a2`](https://github.com/contentful/forma-36/commit/bb6b97a21779bd30bd0894a0c14a7a74eff9c6de) Thanks [@bgutsol](https://github.com/bgutsol)! - Fix Table.Body types
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.41.1
+  - @contentful/f36-typography@4.41.1
+
+## 4.41.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.41.0
+  - @contentful/f36-typography@4.41.0
+
+## 4.40.7
+
+### Patch Changes
+
+- [#2491](https://github.com/contentful/forma-36/pull/2491) [`e36ff8b2`](https://github.com/contentful/forma-36/commit/e36ff8b25a4ee43823b875881aa35b5c03040b49) Thanks [@denkristoffer](https://github.com/denkristoffer)! - fix: differentiate th color from td
+
+* [#2492](https://github.com/contentful/forma-36/pull/2492) [`a2859199`](https://github.com/contentful/forma-36/commit/a28591999695ed0d39080536eff57bb63a8c2842) Thanks [@bgutsol](https://github.com/bgutsol)! - Fix Table.Body children type
+
+* Updated dependencies []:
+  - @contentful/f36-core@4.40.7
+  - @contentful/f36-typography@4.40.7
+
+## 4.40.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.40.6
+  - @contentful/f36-typography@4.40.6
+
+## 4.40.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.40.5
+  - @contentful/f36-typography@4.40.5
+
+## 4.40.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.40.4
+  - @contentful/f36-typography@4.40.4
+
+## 4.40.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.40.3
+  - @contentful/f36-typography@4.40.3
+
+## 4.40.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.40.2
+  - @contentful/f36-typography@4.40.2
+
+## 4.40.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.40.1
+  - @contentful/f36-typography@4.40.1
+
+## 4.40.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.40.0
+  - @contentful/f36-typography@4.40.0
+
+## 4.39.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.39.0
+  - @contentful/f36-typography@4.39.0
+
+## 4.38.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.38.1
+  - @contentful/f36-typography@4.38.1
+
+## 4.38.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.38.0
+  - @contentful/f36-typography@4.38.0
+
+## 4.37.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.37.1
+  - @contentful/f36-typography@4.37.1
+
+## 4.37.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.37.0
+  - @contentful/f36-typography@4.37.0
+
+## 4.36.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.36.0
+  - @contentful/f36-typography@4.36.0
+
 ## 4.35.0
 
 ### Patch Changes